moved tests from filters to actual jinja2 tests
also removed some unused declarations and imports
