+++ lib/ansible/plugins/filter/mathstuff.py
@@ -70,12 +70,6 @@ def max(a):
     _max = __builtins__.get('max')
     return _max(a);
 
-def isnotanumber(x):
-    try:
-        return math.isnan(x)
-    except TypeError:
-        return False
-
 
 def logarithm(x, base=math.e):
     try:
@@ -136,7 +130,6 @@ class FilterModule(object):
     def filters(self):
         return {
             # general math
-            'isnan': isnotanumber,
             'min' : min,
             'max' : max,
